refactor: move footer button components to base gallery module
