+++ src/components/base-gallery/FooterUploadButton.tsx
@@ -1,16 +1,18 @@
-import { FormattedMessage } from "react-intl";
 import { UploadButton } from "../UploadButton";
+import { ReactNode } from "react";
 
 export const FooterUploadButton = ({
   onClick,
+  children,
 }: {
   onClick: (file: File) => void;
+  children: ReactNode;
 }) => (
   <UploadButton
     className="dp-button button-medium ctaTertiary"
     onFile={onClick}
     accept=".jpg, .jpeg, .png, .gif"
   >
-    <FormattedMessage id="upload_image" />
+    {children}
   </UploadButton>
 );
